requirements.txt
streamlit
requests
python-dotenv
google-generativeai
tenacity
fpdf2
